# rowexec: use CheckExistsRequest in LEFT SEMI/ANTI joins