--- a/pkg/sql/rowexec/tablereader.go
+++ b/pkg/sql/rowexec/tablereader.go
@@ -190,7 +190,7 @@ func (tr *tableReader) Start(ctx context.Context) context.Context {
 	if tr.maxTimestampAge == 0 {
 		err = tr.fetcher.StartScan(
 			ctx, tr.FlowCtx.Txn, tr.spans,
-			limitBatches, tr.limitHint, tr.FlowCtx.TraceKV,
+			limitBatches, tr.limitHint, tr.FlowCtx.TraceKV, true,
 		)
 	} else {
 		initialTS := tr.FlowCtx.Txn.ReadTimestamp()