… c-deps/libroach/protos/roachpb/api.pb.h
… pkg/roachpb/api.pb.go
… pkg/roachpb/method_string.go
… pkg/sql/row/fetcher_test.go
… pkg/sql/rowexec/joinreader_test.go
… pkg/sql/rowexec/zigzagjoiner.go
… pkg/sql/tablewriter_delete.go
… pkg/util/cache/cache.go