@helenmhe 2020-09-03
CheckExistsRequest


CheckExistsRequest is currently implemented with a flag checking for types that don’t require right cols, namely left semi/anti joins. Currently in the kv_batch_fetcher, I set  `ba.Header.MaxSpanRequestKeys` and `ba.Header.TargetBytes` to 0, (not limiting the number of rows returned since we aren’t really returning rows). When testing locally it seemed that this had a large impact on performance (parallel performed ~20% better than batch size limited), but I wasn’t able to recreate anything with a low p-value on gceworker.


Testing

Switching to the inner loop in https://github.com/cockroachdb/cockroach/pull/53669/files#diff-2793c01ca6dd699f8ad1e26f24f700dbR1182 seems to have increased variance, compared to before when I was using -count 100 or so to modify b.N, so despite running faster I don’t know if it’s as desirable for getting statistically significant results? Results with gceworker for some reason tended to have slightly more variance, leading to having very high p-values so unfortunately I didn’t get any worthwhile results on the gceworker after making this change. 

The results I got prior to making the benchmark changes are shown below:
```
name                            old time/op    new time/op    delta
LeftSemiJoin/SingleRow/None-12    13.1ms ±13%    12.9ms ± 9%  -1.73%  (p=0.005 n=86+100)

name                            old alloc/op   new alloc/op   delta
LeftSemiJoin/SingleRow/None-12    4.61MB ± 2%    4.57MB ± 1%  -0.87%  (p=0.000 n=97+98)

name                            old allocs/op  new allocs/op  delta
LeftSemiJoin/SingleRow/None-12     28.4k ± 1%     30.3k ± 1%  +6.85%  (p=0.000 n=90+97)
```















Along with the corresponding profile call_trees:

Master

<img width="625" alt="pprof_master" src="https://user-images.githubusercontent.com/26393003/92158783-6debd380-edfa-11ea-98f1-63a4d48cf0fc.png">

CheckExists

<img width="635" alt="pprof_ce" src="https://user-images.githubusercontent.com/26393003/92158804-73491e00-edfa-11ea-811c-9daccde59665.png">


Get

<img width="430" alt="Screen Shot 2020-09-03 at 3 30 16 PM" src="https://user-images.githubusercontent.com/26393003/92158853-85c35780-edfa-11ea-9c7b-b12047572da4.png">


To force the GetRequest code path in BenchmarkLeftSemiJoin I had to create the foo index as UNIQUE.

Overall from the testing, it was a bit difficult to isolate the difference between using Scan vs CheckExists even in a benchmark designed to stress the left semi join. I’m not entirely sure why `fk_test.go` wasn’t hitting this codepath as much, but it could be worthwhile also investigating that.